refactor: structure
* refactor: added controllers

* refactor:: added service

* refactor: added routes

* refactor: added utils
