--- constants.js
@@ -1,3 +1,0 @@
-const BASE_URL = "/api/v1";
-
-export default BASE_URL;
