+++ src/routes/accounts.route.js
@@ -1,0 +1,9 @@
+const express = require("express");
+const router = express.Router();
+const accountsController = require("../controllers/accounts.controller");
+
+router.get("/", accountsController.get);
+
+router.get("/:id", accountsController.getWithId);
+
+module.exports = router;
