+++ src/utils/auth-header.util.js
@@ -1,0 +1,11 @@
+const authHeader = (bearerToken) => {
+    return {
+        headers: {
+            Authorization: `Bearer ${bearerToken}`,
+        },
+    };
+};
+
+module.exports = {
+    authHeader,
+};
